Unit tests - Update: Removed commented out unittest.main(). We can add back later if we need to run all tests in a single file - Bugfix: TEST_DIR wasn't resolving: fixed by adding THIS_TEST_DIR and updating pathing of these variables. Also simplified these path variables. - Bugfix: test_actions_api.py was missing file extension - Delete: enclave_wrangler_tests/test_enclave_wrangler.py: After moving its remaining unmigrated contents to other files. - Delete: Uncommented test_concept_members_enclave_to_db() - Add: empty test_enclave_wrangler.py at root of test directory. This helps solve a tricky merge conflict after failed rebase. - Update: Moved test/input/test_enclave_wrangler into test/test_enclave_wrangler/input - Update: Rename test/enclave_wrangler_tests to test/test_enclave_wrangler. Not a big deal but probably standard convention. - Update: test/test_backend/: Moved test_backend.py there and added __init__ for module imports etc. - Delete: test/__main__.py. This was here to be able to run tests via 'python -m test'. But it was only set up to run one of the files, and I thought maybe this isn't worth maintaining. There are lots of ways to run tests, and we haven't been using this way. What we'll probably mostly use is 'python -m unittest discover' - Update: Code comments: (i) about running tests, (ii) stale todo's - Update: Imports: Removed some stale ones. - Update: Constants: Removed ones not needed in some files. Capitalized one of them per convention.
