--- test/__main__.py
@@ -1,9 +1,0 @@
-"""Tests
-
-TODO's
- - Test framework: Current implementation is ad-hoc for purposes of development.
-"""
-from test.test_backend import runall
-
-
-runall()
